+++ library/src/validations/email/email.ts
@@ -4,17 +4,13 @@ import { getOutput, getPipeIssues } from '../../utils/index.ts';
 /**
  * Creates a validation function that validates an email.
  *
- * Hint: The regex used is not perfect, but should work for most emails.
- *
  * @param error The error message.
  *
  * @returns A validation function.
  */
 export function email<TInput extends string>(error?: ErrorMessage) {
   return (input: TInput): PipeResult<TInput> =>
-    !/^(([^<>()[\].,;:\s@"]+(\.[^<>()[\].,;:\s@"]+)*)|(".+"))@(([^<>()[\].,;:\s@"]+\.)+[^<>()[\].,;:\s@"]{2,})$/i.test(
-      input
-    )
+    !/^[\w+-]+(?:\.[\w+-]+)*@[\da-z]+(?:[.-][\da-z]+)*\.[a-z]{2,}$/i.test(input)
       ? getPipeIssues('email', error || 'Invalid email', input)
       : getOutput(input);
 }