+++ library/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Change regex of `email` validation (pull request #180)
+- Fix types at `brand`, `transform` and `unwrap` method (issue #195)
+
 ## v0.18.0 (September 30, 2023)
 
 - Add `intersection` and `intersectionAsync` schema (pull request #117)