+++ library/CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to the library will be documented in this file.
 ## vX.X.X (Month DD, YYYY)
 
 - Add `notBytes`, `notLength`, `notSize` and `notValue` validation function (pull request #194)
+- Change regex of `email` validation (pull request #180)
+- Fix types at `brand`, `transform` and `unwrap` method (issue #195)
 
 ## v0.18.0 (September 30, 2023)